+++ tiskitpy/compliance/__init__.py
@@ -1,5 +1,5 @@
-from .compliance_noise import ComplianceNoise, to_DBs, from_DBs
-from .compliance_functions import (gravd, calc_norm_compliance, 
-                                   zp_to_norm_compliance)
+from .compliance_noise import ComplianceNoise
+from .compliance import Compliance
 from .psd_vals import PSDVals
 from .earth_model import EarthModel1D
+from .functions import to_DBs, from_DBs