+++ CHANGELOG.md
@@ -139,3 +139,12 @@ Change `MANIFEST.in` to recursively include subfiles/directories of `decimate/`
 - TimeSpans.from_eqs() now takes ``time_bounds`` input instead of ``starttime`` and ``endtime``
 - ``SpectralDensity``, ``DataCleaner`` and ``CleanRotate`` classes take ``remove_eqs``
   and ``avoid_spans`` as input and provide an ``avoided_spans`` property.
+
+## 1.0.1
+Fix bug in ``ResponseFunctions`` plotting
+
+## 1.1
+- Added ``Compliance`` class
+- ``ComplianceNoise`` now gives expected outputs
+- ``SpectralDensity``: ``plots()`` and ``plot_coherences()`` methods now accept objects without identical seed_ids
+- ``EarthModel1D``: Added ``plot()`` and ``calc_ncomp()`` methods